Allow alumni to use UCL API OAuth
In commit a81d6f580f3a3df04cb6839a60e326a63dd14359 an assumption was made that
a lack of the groups header meant that the user was an alumnus. This is not
strictly true: new students can also have a missing group header.
This patch partially reverts the above commit - we still deny access to the
dashboard for alumni (and consequently new students). However, this patch now
allows alumni (and consequenty new students) to successfully log into apps that
user our OAuth system.