Adding migration for changed model
In commit 6a273b30044e0196bc70d5367cf23130d3b2b0d4 a null constraint was added
to the model. However the `makemigrations` command was not run. Hence the change
in the model is not actually live on the database. This patch runs
`makemigrations` and commits the migration file generated.