Do not read state of non-sensor
OccupEye added a "host" as a sensor, but it isn't a sensor. Hence it is not
possible to inspect if it is occupied or not. This patch handles this case by
setting the opacity to 0, hence not misleading anyone into thinking there is a
seat - free or not - at that location.