Fix duplicate labs
In the current timetable for a lot of students all labs would show, but also
with the added caveat that the lab might be mislabelled. This patch gets rid of
the old way to calculate what modules (and groups) they were in and replaces it
with a simpler version that seems to work.