Handle timetable events with no bookings
This patch handles timetable events that have no corresponding bookings
attached to them. This can occur during a "normal" year, but this is
especially prevalent during COVID-19 with online lectures dominating
people's timetables. Interestingly, this was already handled in
_get_timetable_events(), which is still used in some of the Python
only endpoints. Hence, the logic to handle this is adapted into parts
of the SQL and in the serialisation that still happens on the Python
side, where appropriate.