Shib details (#3099)
* Add additional Shibboleth fields to User model

This adds the following additional fields to the database, which will
be filled in from data in the Shibboleth callback:
* mail - long-form email, e.g. fname.lname.year@ucl.ac.uk
* sn - second name.
* affiliation - Not entirely sure...
* unscoped_affiliation - Not entirely sure.

Currently the default value is an empty-string for all fields.

* Fill in additional Shibboleth fields from callback

The Shibboleth callback function used in the OAuth flow now saves the
following additional fields:
mail, sn, affiliation, unscoped_affiliation

This patch also updates the tests to cover this functionality.

* Test creating/updating user in callback

Consequently, also updates cn, which was an oversight!

* Filter User on employee_id instead of eppn (email)

Will avoid having to check for MultipleObjectsReturned exception (as
employee_id is unique in the DB) and will also be faster due to an
index on the User model due to the uniqueness.

* Use convenience method instead of constructor

* Don't reload User object unnecessarily

* Handle Shibboleth responses that violate integrity

It is not impossible that Shibboleth responses could violate the
integrity of our database. This patch adds code to handle this case
gracefully, with corresponding tests.

* Add more details about what attributes are missing

* Add Shibboleth attribute validation function

Adds a function that encapsulates the process of validating the
attributes received from Shibboleth.

* Use same validation function in all callbacks

Switch all callback functions to use validate_shibboleth_callback() for
Shibboleth attribute validation. This patch also includes corresponding
tests for each applicable endpoint

* Update comments

* Sate hound

* Add clarifying comment to model, fold statement

Co-authored-by: Zak Morgan <zcabzsm@ucl.ac.uk>