[LibCal] Add Cancel endpoint
Adds the /space/cancel endpoint, allowing a user to cancel their own
bookings. It is not possible to simply proxy this API, if we were to do
so, someone could easily cancel somebody else's booking. Hence, some
safety checks are done in the view to prevent this scenario.

Tests are also added to confirm the correct functionality of the code.