[LibCal] Check non-JSON response
LibCal sometimes returns non-JSON output (404 usually). Add some code
to make sure we catch this gracefully. We output a 500 as this usually
signifies our regex in urls.py is incorrect or LibCal has done
something odd. In either case, it's unlikely to be the user's fault,
so a 4XX response isn't appropriate IMO.