Specify test runner in mocked settings
Currently the test runner has to be explicitly specified on the command
line. This patch changes this by specifying the runner in the custom
settings file.

The reason for this change is because I couldn't convince PyCharm to
run the tests by specifying the test runner on the command line. This
change allowed me to setup tests directly in PyCharm. In addition to
run tests requires less typing :)