Move AUP acceptance to create-app stage
We can't do it at the dashboard login stage anymore because we are using
the standard OAuth flow, so can't restrict login based on something like
this.

We could move the entire AUP acceptance responsibility to this Next.js
frontend app, but adding an entire database just for that, and
fragmenting our data doesn't seem like a great option.

So now, when a user tries to create an app, the endpoint will return an
error, prompting the frontend to show the AUP, request the user to
accept it, and send the acceptance to a new endpoint.

Then, the user can continue making their app.