Require dashboard APIs to be POST and application/json
We are using SameSite=Lax cookies for the NextAuth.js JWT cookie, which
prevents most types of CSRF attacks.

In the case where a <form method=post> is used on a malicious site, this
cannot be of type 'application/json', so we can protect against such
attacks by ensuring the API is explicitly called with this content type.

Finally, we have disabled CORS on all endpoints, so JS-initiated attacks
also should not work.

SameSite=lax should prevent cross-origin opaque (`no-cors` mode)
requests (where no response is sent) to not be an issue, because the
cookie will not be sent.

As of today, SameSite=Lax support is 95.62% globally
(https://caniuse.com/?search=SameSite).

Useful reading:
- https://michaelzanggl.com/articles/web-security-cors-csrf-samesite/
- https://stackoverflow.com/a/11024387
- https://stackoverflow.com/a/62450640